無限の猿定理を用いたジョーク的な生成方式を実装。 生成方式に関する説明モーダルの実装。
